Add initial Tree Grid PoC implmentation
This commit adds the basic thrust of a tree grid implementation.
It is being made availble in source only, for a very specific
collaboration scenario. It is NOT production ready. The related
refactoring, however, affects both the Tree View and the Tree Grid.

This commit also updates packages, which will resolve some dependabot
PRs.

Documentation is updated to use tree view instead of treeview.